file_search_agent/requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM Providers
openai>=1.0.0
anthropic>=0.18.0
ollama>=0.1.0

# Embeddings
sentence-transformers>=2.2.0
tiktoken>=0.5.0

# Vector Database
faiss-cpu>=1.7.4  # Use faiss-gpu for GPU support
chromadb>=0.4.0

# Document Parsing
PyMuPDF>=1.23.0           # PDF parsing (fitz)
python-docx>=1.1.0        # Word documents
openpyxl>=3.1.0           # Excel files
python-pptx>=0.6.21       # PowerPoint files
pandas>=2.0.0             # CSV/Excel data
chardet>=5.0.0            # Encoding detection
python-magic>=0.4.27      # File type detection

# OCR & Vision
pytesseract>=0.3.10
Pillow>=10.0.0
easyocr>=1.7.0

# File System
watchdog>=3.0.0           # File watching
pathspec>=0.11.0          # Gitignore patterns

# API & CLI
fastapi>=0.109.0
uvicorn>=0.27.0
typer>=0.9.0
rich>=13.0.0              # Beautiful CLI output

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Utilities
tqdm>=4.66.0              # Progress bars
tenacity>=8.2.0           # Retry logic
httpx>=0.26.0             # HTTP client
aiofiles>=23.0.0          # Async file operations

# Optional: Advanced Features
# langchain>=0.1.0        # For advanced RAG
# llama-index>=0.10.0     # Alternative RAG framework
